ovsdb: adjust ovsdb_monitor_changes to hold rows by either transaction id or datum
The next few patchs implements a data structure to track rows according to matched
condition's clauses. The goal is to reduce the computation of conditional monitoring
when all clauses have "==" function (which is the most common case).
This patch adjust ovsdb_monitor_changes to hold rows either since transaction point in
time or hold rows according to a common column value. In the later case we need to
maintain a list of all old values that we have and set the right old value before
using ovsdb_monitor_row.

Signed-off-by: Liran Schour <lirans@il.ibm.com>