ovsdb: Implement data structure to hold changed rows that matches condition's clauses
This patch implements a data structure to hold all changed rows that matches any
condition's clause with function "==" in this ovsdb_monitor_table.
This data structure will be used only for sessions that have only "=="/"includes"
or false function (which is the most common case in conditional monitoring).
We want to limit computation to O(#columns in all clauses) in insertion of new changed row and
O(#sent changes) in composing update message.

Signed-off-by: Liran Schour <lirans@il.ibm.com>