ovsdb: Implement data structure to hold changed rows that matchs condition's clauses
This patch implements a data structure to hold all changed rows that match's any
condition's clauses with function "==" in this ovsdb_monitor_table.
This data structure will be used only for sessions that have maximal
clause function of "==" (which is the most common case in conditional monitoring).
We want to achieve O(#clauses) in insertion of new changed row and O(#sent changes)
in composing update message.

Signed-off-by: Liran Schour <lirans@il.ibm.com>